backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.20
python-dotenv==1.2.1

# Database
motor==3.3.1
pymongo==4.5.0

# Data processing
pandas==2.0.3
numpy==1.26.4
scipy==1.11.4
openpyxl==3.1.5

# Visualization
plotly==5.18.0

# AI/ML
openai==1.12.0
scikit-learn>=1.5.0

# Validation
pydantic==2.12.4
email-validator==2.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Other utilities
python-dateutil==2.9.0.post0
requests==2.32.5
